// src/types/favoriteT.ts
export type favoriteDTO = {
    id?: number;
    usuario_id: number;
    primeiro_jogo_titulo?: string;
    primeiro_jogo_url?: string;
    segundo_jogo_titulo?: string;
    segundo_jogo_url?: string;
    terceiro_jogo_titulo?: string;
    terceiro_jogo_url?: string;
    quarto_jogo_titulo?: string;
    quarto_jogo_url?: string;
    quinto_jogo_titulo?: string;
    quinto_jogo_url?: string;
    data_criacao?: Date;
    data_modificacao?: Date;
};